refactored seed folder structure, added folder for parameter loading
