+++ internal/generate/writer/writer.go
@@ -1,18 +1,20 @@
-package seed
+package writer
 
 import (
 	"encoding/json"
 	"os"
 
 	"github.com/pkg/errors"
+	"github.com/stellar/stellar-rpc-blaster/internal/generate/seed"
+	"github.com/stellar/stellar-rpc-blaster/internal/util"
 )
 
 type OutputSchema struct {
-	LedgerRange Range    `json:"ledger_range"`
-	TxHashes    []TxData `json:"tx_hashes"`
-	ContractIDs []string `json:"contract_ids"`
-	EventTopics []string `json:"event_topics"`
-	LedgerKeys  []string `json:"ledger_keys"`
+	LedgerRange util.Range    `json:"ledger_range"`
+	TxHashes    []seed.TxData `json:"tx_hashes"`
+	ContractIDs []string      `json:"contract_ids"`
+	EventTopics []string      `json:"event_topics"`
+	LedgerKeys  []string      `json:"ledger_keys"`
 }
 
 // SeedWriter accumulates seed data into an OutputSchema struct,
@@ -41,27 +43,4 @@ func (w *SeedWriter) Flush() error {
 		return errors.Wrap(err, "failed to encode output")
 	}
 	return nil
-}
-
-// Slice returns the accumulated items from an Entry.
-func (e *Entry[T]) Slice() []T {
-	return e.Map[e.key].([]T)
-}
-
-// buffer for accumulating items under a JSON key
-type Entry[T any] struct {
-	Map map[string]any
-	key string
-}
-
-// NewEntry creates a new write buffer with the given JSON key and element type
-func NewEntry[T any](key string, cap int) Entry[T] {
-	m := make(map[string]any)
-	m[key] = make([]T, 0, cap)
-	return Entry[T]{Map: m, key: key}
-}
-
-// Append adds an item to the entry's slice
-func (e *Entry[T]) Append(item T) {
-	e.Map[e.key] = append(e.Map[e.key].([]T), item)
 }
