+++ internal/run/parameters/endpoints.go
@@ -1,0 +1,1 @@
+package parameters
